# Cache labels.Labels to Identify Series in Remote Write