@brian-brazil 2019-08-13
That's looking okay. What do the benchmarks say now?