@csmarchbanks 2019-08-13
I removed the merge conflict, and updated the benchmarks, slightly fewer allocations than last time, CPU jumps around between runs but is about the same before/after.

I don't expect the graphs to change too much, but can repost them in a few hours if interested.